+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## [Unreleased](https://github.com/mertowitch/PHPneeds/compare/v0.3.2...HEAD)
-- /images directory moved to [/public/assets/images](/public/assets/images)
+- **/images** directory moved to [/public/assets/images](/public/assets/images)
 - Some edits on;
   - README.md
   - CHANGELOG.md