+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased](https://github.com/mertowitch/PHPneeds/compare/v0.3.2...HEAD)
+### Changed
 - **/images** directory moved to [/public/assets/images](/public/assets/images)
 - Some edits on;
   - README.md